pyproject.toml
[project]
name = "wework-robot-opencode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28.0",
    "flask>=2.3.0",
    "pycryptodome>=3.20.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
